# openapi.yaml
openapi: 3.0.3
info:
  title: 协同管理工具 API
  version: 0.1.0
  description: |
    Spec-first 唯一契约源。
    Go 侧用 oapi-codegen 生成 server 接口与模型，前端用 openapi-typescript + openapi-fetch 生成类型化客户端；生成物不手改。
    业务规则只在后端 domain 包实现，界面反馈一律消费本契约返回的派生字段。
servers:
  - url: /api/v1
...
paths:
...
  /projects/{projectId}/members:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      operationId: listProjectMembers
      tags: [projects]
      summary: 项目成员列表（含成员角色）
      responses:
        "200":
          description: 成员列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectMember"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: addProjectMembers
      tags: [projects]
      summary: 将一名或多名用户加入项目并赋予同一成员角色（仅项目管理员／项目负责人）
      description: |
        一次提交建立全部成员关系并返回逐人结果：已在项目内或用户不存在的按人跳过，
        不牵连名单里的其余人；名单为空或角色不合法整批拒绝。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddProjectMembersRequest"
      responses:
        "201":
          description: 已处理（逐人结果见响应体，added 可能为空）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddProjectMembersResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /projects/{projectId}/tasks/{taskId}/member-inputs:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: taskId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    post:
      operationId: createMemberInput
      tags: [tasks]
      summary: 指定项目成员提供输入（AC-29；可一次多选对接人，AC-53）：为每名对接人建立「成员 → 目标任务」交付物边并生成输入请求；输入源属关键字段，进所属 KR 负责人审批（AC-23），输入关系建立后即发站内通知（裁决 #162）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMemberInputRequest"
      responses:
        "200":
          description: 已受理，返回任务最新状态（进入审批时边与输入请求尚未建立，任务上带待审批变更单）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /projects/{projectId}/input-requests/{requestId}/accept:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: requestId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    post:
      operationId: acceptInputRequest
      tags: [tasks]
      summary: 对接人同意接收（AC-30）；接收只表示承担责任，不代表输入就绪
      responses:
        "200":
          description: 已接收
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InputRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /projects/{projectId}/input-requests/{requestId}/provide:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: requestId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    post:
      operationId: provideInput
      tags: [tasks]
      summary: 对接人提交内容或文件（AC-30）；提交后输入请求变为已提供，目标任务输入更新为就绪
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProvideInputRequest"
      responses:
        "200":
          description: 已提交
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvideInputResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /projects/{projectId}/input-requests/{requestId}/commit:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: requestId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    post:
      operationId: commitInputRequestFile
      tags: [inputs]
      summary: 确认输入附件已写入对象存储（两阶段提交第二步；校验通过后输入才转为已提供）
      responses:
        "200":
          description: 已确认，输入转为已提供
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InputRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /projects/{projectId}/input-requests/{requestId}/file-url:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: requestId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      operationId: getInputRequestFileUrl
      tags: [tasks]
      summary: 输入请求已提交文件的预签名下载地址
      responses:
        "200":
          description: 预签名地址
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadUrlResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /projects/{projectId}/report/export:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: range
        in: query
        required: false
        schema:
          $ref: "#/components/schemas/ReportRange"
      - name: format
        in: query
        required: true
        schema:
          type: string
          enum: [pdf, image]
          description: pdf＝正式留档 PDF；image＝移动端长图 PNG
    get:
      operationId: exportReport
      tags: [tasks]
      summary: 导出报告（AC-20）：经 Gotenberg 渲染为正式 PDF 或移动端长图
      responses:
        "200":
          description: 导出文件
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "502":
          description: 渲染服务不可用
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  responses:
    Unauthorized:
      description: 未登录或会话已过期
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ValidationError:
      description: 请求内容未通过校验
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: 当前用户无权执行该动作（动作权限按成员角色在 domain 层统一判定）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: 与现有资源状态冲突
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: 机器可读错误码
        message:
          type: string
          description: 面向用户的错误说明
...
    MemberRole:
      type: string
      description: 成员角色（项目管理员／项目成员／访客），见词汇表「成员角色」
      enum: [admin, member, viewer]
    ProjectMember:
      type: object
      required: [userId, username, displayName, role]
      properties:
        userId:
          type: integer
          format: int64
        username:
          type: string
        displayName:
          type: string
        role:
          $ref: "#/components/schemas/MemberRole"
        roleLabel:
          type: string
          description: 成员角色显示文案（派生字段；前端不按枚举拼文案）
    AddProjectMembersRequest:
      type: object
      required: [userIds, role]
      properties:
        userIds:
          type: array
          minItems: 1
          description: 本次加入的用户 id 名单，全部赋予同一个成员角色；重复的 id 只算一次
          items:
            type: integer
            format: int64
        role:
          $ref: "#/components/schemas/MemberRole"
    AddProjectMembersResult:
      type: object
      required: [added, skipped]
      description: 一次批量加入的逐人结果
      properties:
        added:
          type: array
          description: 本次新建立成员关系的人
          items:
            $ref: "#/components/schemas/ProjectMember"
        skipped:
          type: array
          description: 未加入的人及原因
          items:
            $ref: "#/components/schemas/SkippedMember"
    SkippedMember:
      type: object
      required: [userId, reason, reasonLabel]
      properties:
        userId:
          type: integer
          format: int64
        displayName:
          type: string
          description: 用户姓名；用户不存在时缺省
        reason:
          type: string
          description: 跳过原因枚举
          enum: [already_member, user_not_found]
        reasonLabel:
          type: string
          description: 跳过原因显示文案（派生字段；前端不按枚举拼文案）
...
    TaskStatus:
      type: string
      description: 任务生命周期状态（词汇表「任务生命周期状态」）；页面主状态汇总，审批原始状态在审批单中
      enum:
        [
          not_started,
          waiting_input,
          in_progress,
          pending_intermediate_review,
          pending_final_review,
          completed,
          cancelled,
        ]
    ResultUpdateState:
      type: string
      description: >-
        成果更新进程（词汇表「成果更新」；派生自任务事实，接口不接受直接写入）：
        none＝无；open＝已发起、候选内容尚未随完成申请提交；reviewing＝已提交，或签或 KR 终审在审。
        整个过程中任务生命周期状态保持 completed
      enum: [none, open, reviewing]
    Task:
      type: object
      required:
        [
          id,
          keyResultId,
          code,
          name,
          ownerId,
          ownerName,
          startDate,
          endDate,
          updatedAt,
          status,
          statusLabel,
          currentStage,
          canProposeFieldChange,
          canStart,
          canUpdateProgress,
          canCancel,
          receiverScope,
        ]
      properties:
        id:
          type: integer
          format: int64
        keyResultId:
          type: integer
          format: int64
        code:
          type: string
          description: 任务展示编号，形如 T1.1.1（AC-64；创建时分配并持久保存，删除同级任务后不重排）
        name:
          type: string
        ownerId:
          type: integer
          format: int64
          description: 任务负责人
        ownerName:
          type: string
          description: 任务负责人姓名（派生字段）
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        updatedAt:
          type: string
          format: date-time
          description: 任务最后更新时间（任务详情页头展示，AC-50）
        status:
          $ref: "#/components/schemas/TaskStatus"
        statusLabel:
          type: string
          description: 面向用户的状态显示文案（AC-04、决策 34；派生字段）；审批等待状态为“待{当前审批人姓名}审批”，或签多人为“待{首位姓名}等N人审批”，无审批人时退化为“待审批”
        description:
          type: string
          description: 任务说明，选填
        completionCriteria:
          type: string
          description: 完成标准，选填
        currentStage:
          type: string
          description: 当前环节（词汇表「当前环节」；中文标签，派生字段）
        pendingActorId:
          type: integer
          format: int64
          description: 待行动人（词汇表「待行动人」；派生字段，无待行动人时不返回）
        pendingActorName:
          type: string
        progress:
          type: integer
          minimum: 0
          maximum: 100
          description: 可选进度百分比（词汇表「任务进度」）；已完成任务一律 100 且不可编辑，未填写时不返回、前端只展示状态（AC-63）
        cancelReason:
          type: string
          description: 取消原因（已取消任务保留，PRD §5.1）
        canStart:
          type: boolean
          description: 当前用户能否开始执行本任务（派生字段；负责人／可编辑项目者，未开始或等待输入时）
        canUpdateProgress:
          type: boolean
          description: 当前用户能否更新本任务进度（派生字段；负责人／可编辑项目者，仅进行中；完成后锁定为 100，AC-63）
        canCancel:
          type: boolean
          description: 当前用户能否发起取消申请（派生字段；任务负责人／项目管理员，非终态且任务上没有未决审批单，AC-57）
        fieldChange:
          $ref: "#/components/schemas/FieldChange"
          description: 最近一张需要关注的变更单（待审批＝拟议值标示，或退回未处理＝退回待处理事项）；无则不返回
        deliverableNames:
          type: array
          description: 交付物项名称列表（列表「预期交付物」列展示用，派生字段）
          items:
            type: string
        canProposeFieldChange:
          type: boolean
          description: 当前用户能否编辑任务／提交关键字段修改（派生字段）
        fieldEditMode:
          type: string
          enum: [exempt, approval]
          description: >-
            就地编辑的保存路由（#138 裁决 E1；派生自 FieldChangeRoute，前端不复算规则）：
            exempt=KR 负责人本人免审即时生效；approval=生成变更单进入审批
            （保存时须填修改原因）。不可编辑时不返回（以 canProposeFieldChange 为准）
        pendingReviewCount:
          type: integer
          description: 本任务上未决审批单条数（派生字段；关键字段变更／取消、完成申请合计，前端不再自行相加）
        canDiscuss:
          type: boolean
          description: 当前用户能否提交任务讨论（派生字段；全体项目内成员含访客，公开项目的隐式访客不可，见词汇表「隐式访客」）
        canManageDeliverables:
          type: boolean
          description: 当前用户能否配置任务交付物项（派生字段；负责人／创建人／可编辑项目者，非终态）
        canSubmitCompletion:
          type: boolean
          description: 当前用户能否提交完成申请（派生字段；负责人，进行中且有候选内容）
        canManageReviewers:
          type: boolean
          description: 当前用户能否调整成果审核人配置（派生字段；负责人／创建人／可编辑项目者，审核中与终态不可）
        openBlockerCount:
          type: integer
          description: 当前派生卡点数量（派生字段）
        canUploadCandidate:
          type: boolean
          description: 当前用户能否登记候选交付物内容（派生字段；负责人，执行类状态）
        receiverScope:
          $ref: "#/components/schemas/ReceiverScope"
        receivers:
          type: array
          description: 指定成员为接收方时的名单（receiverScope=members；其余取值为空数组）
          items:
            $ref: "#/components/schemas/ReviewerInfo"
        canManageReceivers:
          type: boolean
          description: 当前用户能否配置接收方（派生字段；与输入／输出配置同口径：负责人／创建人／可编辑项目者，终态不可）
        canConfirmReceipt:
          type: boolean
          description: 当前用户是否有本任务的待接收项待确认（派生字段；MW-09）
        participants:
          type: array
          description: >-
            参与人名单（词汇表「参与人」；PRD §9.2 按需字段）：任务上除负责人以外的协作者，
            只作展示与检索——不产生待办、不进审批链、不影响权限、不参与我的工作归组与排序。
            未配置时为空数组
          items:
            $ref: "#/components/schemas/ReviewerInfo"
        canManageParticipants:
          type: boolean
          description: 当前用户能否配置参与人（派生字段；负责人／创建人／可编辑项目者，终态不可）
        resultUpdate:
          $ref: "#/components/schemas/ResultUpdateState"
        canStartResultUpdate:
          type: boolean
          description: >-
            当前用户能否对本任务发起成果更新（派生字段；AC-66：任务负责人／项目管理员，
            任务已完成、无在途成果更新且没有其他未决审批单）
...
    DownloadUrlResponse:
      type: object
      required: [url]
      properties:
        url:
          type: string
          description: MinIO 预签名 GET 地址（预览/下载）
...
    Necessity:
      type: string
      description: 输入必要性（必要／参考）
      enum: [required, reference]
...
    ReportRange:
      type: string
      description: 报告时间范围（PRD §7.8）
      enum: [today, week, month, all]
...
    InputRequestState:
      type: string
      description: 输入请求状态（词汇表「输入请求」；PRD §5.5）。uploading：已登记待上传，文件确认写入后才转 provided
      enum: [pending, accepted, provided, uploading]
    InputRequest:
      type: object
      description: 指定项目成员的输入请求（附着在成员 → 目标任务的交付物边上）
      required: [id, edgeId, providerId, providerName, state]
      properties:
        id:
          type: integer
          format: int64
        edgeId:
          type: integer
          format: int64
        providerId:
          type: integer
          format: int64
        providerName:
          type: string
        contentNote:
          type: string
          description: 所需内容说明
        state:
          $ref: "#/components/schemas/InputRequestState"
        notifiedAt:
          type: string
          format: date-time
          description: 站内通知发送时间（裁决 #162：输入关系建立后即发送）
        acceptedAt:
          type: string
          format: date-time
        providedAt:
          type: string
          format: date-time
        providedText:
          type: string
          description: 对接人提交的文字内容
        providedFileName:
          type: string
        canAccept:
          type: boolean
          description: 当前用户能否同意接收（派生字段；对接人本人且待接收）
        canProvide:
          type: boolean
          description: 当前用户能否提交内容（派生字段；对接人本人且已接收）
    CreateMemberInputRequest:
      type: object
      description: |
        新增输入要求（来源＝指定项目成员，AC-29；可一次多选对接人，AC-53）。
        不收输入名称：输入源的可读标识取「所需内容」摘要（裁决 F1），见 DeliverableEdge.name。
      required: [necessity, providerIds, contentNote, expectedDate]
      properties:
        necessity:
          $ref: "#/components/schemas/Necessity"
        providerIds:
          type: array
          minItems: 1
          items:
            type: integer
            format: int64
          description: 对接人（非只读项目成员，AC-53 多选）；每人分别建边并生成输入请求与通知，不可重复
        contentNote:
          type: string
          minLength: 1
          maxLength: 500
          description: 所需内容（§9.1 必填）
        expectedDate:
          type: string
          format: date
          description: 期望时间（§9.1 必填）
    ProvideInputRequest:
      type: object
      properties:
        text:
          type: string
          maxLength: 2000
          description: 文字内容；与文件至少提交其一
        fileName:
          type: string
          maxLength: 200
          description: 上传文件名；返回预签名上传地址
    ProvideInputResponse:
      type: object
      required: [request]
      properties:
        request:
          $ref: "#/components/schemas/InputRequest"
        uploadUrl:
          type: string
          description: 提交了 fileName 时返回的预签名上传地址
...
    ReviewerInfo:
      type: object
      required: [userId, displayName]
      properties:
        userId:
          type: integer
          format: int64
        displayName:
          type: string
    ReceiverScope:
      type: string
      description: 接收方范围（词汇表「接收方」；模块 PRD §8.6）：不配置、指定成员或所有项目成员
      enum: [none, members, all]
...
security:
  - cookieAuth: []